// UnsDashboard/unsdashboard.client/src/opcua/models/DataSetWriterDataType.ts
export interface DataSetWriterDataType {
    /**
     * 
     * @type {string}
     * @memberof DataSetWriterDataType
     */
    Name?: string;
    /**
     * 
     * @type {boolean}
     * @memberof DataSetWriterDataType
     */
    Enabled?: boolean;
    /**
     * 
     * @type {number}
     * @memberof DataSetWriterDataType
     */
    DataSetWriterId?: number;
    /**
     * 
     * @type {number}
     * @memberof DataSetWriterDataType
     */
    DataSetFieldContentMask?: number;
    /**
     * 
     * @type {number}
     * @memberof DataSetWriterDataType
     */
    KeyFrameCount?: number;
    /**
     * 
     * @type {string}
     * @memberof DataSetWriterDataType
     */
    DataSetName?: string;
    /**
     * 
     * @type {Array<KeyValuePair>}
     * @memberof DataSetWriterDataType
     */
    DataSetWriterProperties?: Array<KeyValuePair>;
    /**
     * 
     * @type {ExtensionObject}
     * @memberof DataSetWriterDataType
     */
    TransportSettings?: ExtensionObject;
    /**
     * 
     * @type {ExtensionObject}
     * @memberof DataSetWriterDataType
     */
    MessageSettings?: ExtensionObject;
}