// UnsDashboard/unsdashboard.client/src/opcua/models/BrokerWriterGroupTransportDataType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OPC UA REST API
 * This API provides simple REST based access to an OPC UA server.
 *
 * The version of the OpenAPI document: 0.0.1
 * Contact: office@opcfoundation.org
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BrokerWriterGroupTransportDataType
 */
export interface BrokerWriterGroupTransportDataType {
    /**
     * 
     * @type {string}
     * @memberof BrokerWriterGroupTransportDataType
     */
    QueueName?: string;
    /**
     * 
     * @type {string}
     * @memberof BrokerWriterGroupTransportDataType
     */
    ResourceUri?: string;
    /**
     * 
     * @type {string}
     * @memberof BrokerWriterGroupTransportDataType
     */
    AuthenticationProfileUri?: string;
    /**
     * 
     * @type {number}
     * @memberof BrokerWriterGroupTransportDataType
     */
    RequestedDeliveryGuarantee?: number;
}

/**
 * Check if a given object implements the BrokerWriterGroupTransportDataType interface.
 */
export function instanceOfBrokerWriterGroupTransportDataType(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function BrokerWriterGroupTransportDataTypeFromJSON(json: any): BrokerWriterGroupTransportDataType {
    return BrokerWriterGroupTransportDataTypeFromJSONTyped(json, false);
}

export function BrokerWriterGroupTransportDataTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): BrokerWriterGroupTransportDataType {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'QueueName': !exists(json, 'QueueName') ? undefined : json['QueueName'],
        'ResourceUri': !exists(json, 'ResourceUri') ? undefined : json['ResourceUri'],
        'AuthenticationProfileUri': !exists(json, 'AuthenticationProfileUri') ? undefined : json['AuthenticationProfileUri'],
        'RequestedDeliveryGuarantee': !exists(json, 'RequestedDeliveryGuarantee') ? undefined : json['RequestedDeliveryGuarantee'],
    };
}

export function BrokerWriterGroupTransportDataTypeToJSON(value?: BrokerWriterGroupTransportDataType | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'QueueName': value.QueueName,
        'ResourceUri': value.ResourceUri,
        'AuthenticationProfileUri': value.AuthenticationProfileUri,
        'RequestedDeliveryGuarantee': value.RequestedDeliveryGuarantee,
    };
}